Fixes need int not float.

+++ rsCNN/evaluation/shared/subplots.py
@@ -220,7 +220,7 @@ def _add_internal_window_to_subplot(sampled: samples.Samples, ax: plt.Axes) -> N
     internal_window_radius = sampled.network_config['architecture']['internal_window_radius']
     if (internal_window_radius * 2 == inshape[0]):
         return
-    buffer = (inshape[0] - internal_window_radius * 2) / 2
+    buffer = int((inshape[0] - internal_window_radius * 2) / 2)
     rect = patches.Rectangle(
         (buffer, buffer), internal_window_radius * 2, internal_window_radius * 2, linewidth=1, edgecolor='white',
         facecolor='none'
